move create contents function to content.go
